6/20

+++ site/templates/content/authorizenet/server-process-request.php
@@ -2,7 +2,7 @@
 	header('Content-Type: application/json');
 	$gateway = $page->name; 
 	$ordn = $input->get->text('ordn');
-
+	
 	if ($user->isLoggedin()) {
 		
 	} elseif ($input->post->login && $input->post->password) {
@@ -12,6 +12,7 @@
 	    $user = $session->login($username, $pass);
 		
 	    if ($user) {
+			
 	    } else {
 	    	$response = array(
 				'error' => true,
@@ -26,16 +27,24 @@
 	}
 
 	if ($user) {
-		
 		if ($input->get->debug) {
-			$postrequest = json_decode(file_get_contents($config->paths->content.'test/request.json'));
+			if ($input->get->text('debug') == 'record') {
+				$postrequest = get_request_record($ordn, false);
+				unset($postrequest['cc']); unset($postrequest['expiredate']); unset($postrequest['cvc']); 
+				unset($postrequest['track1']); unset($postrequest['track2']);
+			} else {
+				$postrequest = json_decode(file_get_contents($config->paths->content.'test/request.json'));
+			}
 		} else {
 			$postrequest = $input->post->request;
 		}
 		delete_authorize_responserecord($ordn);
-
-		delete_authorize_record($ordn);
-		writeauthnetrecord($postrequest, false);
+		if (!$input->get->text('debug')) {
+			delete_authorize_record($ordn);
+			writeauthnetrecord($postrequest, false);
+		}
+		
+		
 		include ($config->paths->content.$gateway."/$gateway-router.php");
 	} else {
 		echo json_encode(array('response' => $response));
